requirements.txt
async-exit-stack==1.0.1
async-generator==1.10
fastapi==0.58.1
fasteners==0.14.1
fastjsonschema==2.15.3
Flask==2.1.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
huggingface-hub==0.9.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
numpy==1.22.4
pandas==1.4.2
swagger-ui-bundle==0.0.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.12.1
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
transformers==4.22.0
uvicorn==0.19.0
uvloop==0.17.0
python-dotenv==0.21.0
pymongo==4.1.1
